chore(api): drop RomM < 4.7 support, require minimum server version
Consolidate RommApiV46 + RommApiV47 + ApiRouter into single RommApi
adapter. Remove version routing, capabilities system, and all v4.6
fallback paths. Hard version gate rejects servers below 4.7.0 with
full error page replacement in QAM and game detail UI.
