+++ src/components/RomMGameInfoPanel.tsx
@@ -30,18 +30,17 @@ import {
   getAchievementProgress,
   getSaveSlots,
   isSaveTrackingConfigured,
-  getServerCapabilities,
   debugLog,
   refreshMigrationState,
   logError,
 } from "../api/backend";
-import type { ServerCapabilities } from "../api/backend";
 import { SlotSetupWizard } from "./SlotSetupWizard";
 import { SavesTab } from "./SavesTab";
 import type { RomMetadata, InstalledRom, BiosStatus, SaveStatus, PendingConflict, Achievement, AchievementProgress, EarnedAchievement, SaveSlotSummary } from "../types";
 import { getMigrationState, onMigrationChange, setMigrationStatus } from "../utils/migrationStore";
 import { getSaveSortMigrationState, onSaveSortMigrationChange, setSaveSortMigrationStatus } from "../utils/saveSortMigrationStore";
 import { scrollFocusedToCenter } from "../utils/scrollHelpers";
+import { getVersionError } from "../utils/connectionState";
 
 interface RomMGameInfoPanelProps {
   appId: number;
@@ -71,7 +70,6 @@ interface PanelState {
   activeSlot: string | null;
   availableSlots: SaveSlotSummary[];
   slotsLoading: boolean;
-  capabilities: ServerCapabilities;
 }
 
 /** Format a Unix timestamp (seconds) as a release date string (e.g. "15 Mar 2003") */
@@ -105,13 +103,6 @@ function refreshSlotState(
 }
 
 export const RomMGameInfoPanel: FC<RomMGameInfoPanelProps> = ({ appId }) => {
-  const defaultCapabilities: ServerCapabilities = {
-    device_sync: false,
-    version_history: false,
-    slot_deletion: false,
-    device_management: false,
-  };
-
   const [state, setState] = useState<PanelState>({
     loading: true,
     romId: null,
@@ -136,7 +127,6 @@ export const RomMGameInfoPanel: FC<RomMGameInfoPanelProps> = ({ appId }) => {
     activeSlot: "default",
     availableSlots: [],
     slotsLoading: false,
-    capabilities: defaultCapabilities,
   });
   const romIdRef = useRef<number | null>(null);
   const [migrationPending, setMigrationPending] = useState(getMigrationState().pending);
@@ -242,7 +232,6 @@ export const RomMGameInfoPanel: FC<RomMGameInfoPanelProps> = ({ appId }) => {
           activeSlot: "default",
           availableSlots: [],
           slotsLoading: false,
-          capabilities: defaultCapabilities,
         });
 
         // Check if save slot tracking is configured for this game
@@ -286,17 +275,7 @@ export const RomMGameInfoPanel: FC<RomMGameInfoPanelProps> = ({ appId }) => {
           );
         }
 
-        // Server capabilities (for save sync feature gating)
-        let fetchedCaps: ServerCapabilities | null = null;
-        if (cached.save_sync_enabled) {
-          bgPromises.push(getServerCapabilities().then((c) => { fetchedCaps = c; }).catch(() => {}));
-        }
-
         await Promise.all(bgPromises);
-
-        if (!cancelled && fetchedCaps) {
-          setState((prev) => ({ ...prev, capabilities: fetchedCaps! }));
-        }
       } catch (e) {
         debugLog(`RomMGameInfoPanel: loadData error: ${e}`);
         if (!cancelled) setState((prev) => ({ ...prev, loading: false, error: true }));
@@ -488,6 +467,19 @@ export const RomMGameInfoPanel: FC<RomMGameInfoPanelProps> = ({ appId }) => {
       title ? createElement("div", { className: "romm-panel-section-title" }, title) : null,
       ...children.filter(Boolean),
     );
+
+  // --- Version mismatch — replace entire panel ---
+  const versionError = getVersionError();
+  if (versionError) {
+    return createElement("div", {
+      "data-romm": "true",
+      style: { padding: "16px 2.8vw" },
+    },
+      createElement("div", {
+        style: { fontSize: "13px", color: "rgba(255, 255, 255, 0.7)" },
+      }, versionError),
+    );
+  }
 
   // --- Loading state ---
   // Use minHeight so Steam's scroll container allocates enough space
@@ -1037,7 +1029,6 @@ export const RomMGameInfoPanel: FC<RomMGameInfoPanelProps> = ({ appId }) => {
         activeSlot: state.activeSlot,
         availableSlots: state.availableSlots,
         slotsLoading: state.slotsLoading,
-        capabilities: state.capabilities,
         onSlotSwitched: (newSlot, newStatus) => {
           setState((prev) => ({
             ...prev,
