+++ tests/domain/test_save_sync.py
@@ -86,93 +86,6 @@ class TestDetermineSyncAction:
         )
         assert result == "conflict"
 
-    # --- v4.6 fallback (device_sync_info=None, file_state provided) ---
-
-    def test_v46_fallback_server_unchanged_local_unchanged_skips(self):
-        server_save = {"id": 1, "updated_at": "2026-02-17T06:00:00Z", "file_size_bytes": 1024}
-        file_state = {
-            "last_sync_server_updated_at": "2026-02-17T06:00:00Z",
-            "last_sync_server_size": 1024,
-        }
-        result = determine_sync_action(
-            local_changed=False,
-            server_save=server_save,
-            device_sync_info=None,
-            file_state=file_state,
-        )
-        assert result == "skip"
-
-    def test_v46_fallback_server_unchanged_local_changed_uploads(self):
-        server_save = {"id": 1, "updated_at": "2026-02-17T06:00:00Z", "file_size_bytes": 1024}
-        file_state = {
-            "last_sync_server_updated_at": "2026-02-17T06:00:00Z",
-            "last_sync_server_size": 1024,
-        }
-        result = determine_sync_action(
-            local_changed=True,
-            server_save=server_save,
-            device_sync_info=None,
-            file_state=file_state,
-        )
-        assert result == "upload"
-
-    def test_v46_fallback_server_changed_local_unchanged_downloads(self):
-        server_save = {"id": 1, "updated_at": "2026-02-17T08:00:00Z", "file_size_bytes": 2048}
-        file_state = {
-            "last_sync_server_updated_at": "2026-02-17T06:00:00Z",
-            "last_sync_server_size": 1024,
-        }
-        result = determine_sync_action(
-            local_changed=False,
-            server_save=server_save,
-            device_sync_info=None,
-            file_state=file_state,
-        )
-        assert result == "download"
-
-    def test_v46_fallback_both_changed_conflicts(self):
-        server_save = {"id": 1, "updated_at": "2026-02-17T08:00:00Z", "file_size_bytes": 2048}
-        file_state = {
-            "last_sync_server_updated_at": "2026-02-17T06:00:00Z",
-            "last_sync_server_size": 1024,
-        }
-        result = determine_sync_action(
-            local_changed=True,
-            server_save=server_save,
-            device_sync_info=None,
-            file_state=file_state,
-        )
-        assert result == "conflict"
-
-    # --- v4.6 indeterminate fallback (check_server_changes_fast returns None) ---
-
-    def test_v46_indeterminate_local_also_changed_conflicts(self):
-        """check_server_changes_fast returns None (timestamp changed) → assume server changed.
-        Both local and server changed → conflict."""
-        server_save = {"id": 1, "updated_at": "2026-02-17T10:00:00Z", "file_size_bytes": 1024}
-        # Timestamp changed → indeterminate
-        file_state = {"last_sync_server_updated_at": "2026-02-17T06:00:00Z"}
-        result = determine_sync_action(
-            local_changed=True,
-            server_save=server_save,
-            device_sync_info=None,
-            file_state=file_state,
-        )
-        assert result == "conflict"
-
-    def test_v46_indeterminate_only_server_changed_downloads(self):
-        """check_server_changes_fast returns None → assume server changed.
-        Only server changed → download."""
-        server_save = {"id": 1, "updated_at": "2026-02-17T10:00:00Z", "file_size_bytes": 1024}
-        file_state = {"last_sync_server_updated_at": "2026-02-17T06:00:00Z"}
-        result = determine_sync_action(
-            local_changed=False,
-            server_save=server_save,
-            device_sync_info=None,
-            file_state=file_state,
-        )
-        assert result == "download"
-
     # --- Safe default: server_save exists, no device_sync_info, no file_state ---
 
     def test_server_exists_no_sync_info_no_file_state_assumes_server_changed(self):
